CSCMETAX-185: [REF] Reorganize tests. Renamed and moved some files, detailed re-org to CatalogRecord tests
