+++ src/metax_api/tests/api/base/apitests/authnz/__init__.py
@@ -1,0 +1,1 @@
+from .authnz import ApiAuthnzTestV1
