+++ src/metax_api/tests/api/base/apitests/catalog_records/__init__.py
@@ -1,0 +1,2 @@
+from .read import *
+from .write import *
