+++ src/metax_api/tests/api/base/apitests/contracts/__init__.py
@@ -1,0 +1,1 @@
+from .contracts import ContractApiReadTestV1, ContractApiWriteTestV1
