+++ src/metax_api/tests/api/base/apitests/data_catalogs/__init__.py
@@ -1,0 +1,1 @@
+from .data_catalogs import DataCatalogApiReadTestV1, DataCatalogApiWriteTestV1
