+++ src/metax_api/tests/api/base/apitests/files/__init__.py
@@ -1,0 +1,1 @@
+from .files import FileApiReadTestV1, FileApiWriteTestV1
