--- src/metax_api/tests/api/base/views/__init__.py
@@ -1,1 +1,0 @@
-from .test_file_view import FileViewSetTests
