+++ src/metax_api/tests/models/__init__.py
@@ -1,4 +1,3 @@
-from .test_data_catalog_model import DataCatalogModelTests
-from .test_file_model import FileModelBasicTest, FileManagerTests
-from .test_catalog_record_model import CatalogRecordModelBasicTest, CatalogRecordModelTests, CatalogRecordManagerTests
-from ...services.common_service import CommonService
+from .data_catalog import DataCatalogModelTests
+from .file import FileModelBasicTest, FileManagerTests
+from .catalog_record import CatalogRecordModelBasicTest, CatalogRecordModelTests, CatalogRecordManagerTests
